backend/requirements.txt
requests
Flask
Flask-SocketIO 
flask-cors
eventlet
pandas
python-dotenv